feat: add LangDirProvider component
